Delete Option added
I added the option to delete a row that the user selected. Currently the problems are that the values from bellow arent copied to the upper row and that the values of the selected row arent getting deleted.
the Bug with the Textbox will be fixed after i have eveything else done since it doesn't need a fix fast and that i don't know how this bug even exists and how to fix it.